refactor: 우승자 관련 로직 JudgeService로 분리
