+++ src/main/java/Application.java
@@ -1,9 +1,8 @@
 import controller.RacingCarController;
+import service.JudgeService;
 import service.RacingCarService;
 import view.InputView;
 import view.OutputView;
-
-import java.util.Scanner;
 
 public class Application {
 
@@ -11,8 +10,9 @@ public class Application {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
         RacingCarService racingCarService = new RacingCarService();
+        JudgeService judgeService = new JudgeService();
 
-        RacingCarController racingCarController = new RacingCarController(racingCarService, inputView, outputView);
+        RacingCarController racingCarController = new RacingCarController(racingCarService, inputView, outputView, judgeService);
 
         racingCarController.run();
     }
