+++ src/test/java/service/RacingCarServiceTest.java
@@ -1,18 +1,13 @@
 package service;
 
-import domain.Cars;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
-import static util.Constants.MIN_FORWARD_NUMBER;
 
 class RacingCarServiceTest {
     private RacingCarService racingCarService;
@@ -40,47 +35,4 @@ class RacingCarServiceTest {
     void getCountInvalidNumberTest(String count) {
         assertThatThrownBy(() -> racingCarService.getCount(count)).isInstanceOf(IllegalArgumentException.class);
     }
-
-    @DisplayName("우승자 1명 테스트")
-    @Test
-    void findWinnerTest() {
-        Cars cars = new Cars(List.of("가", "나", "다"));
-
-        cars.getCars().get(0).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(1).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(1).drive(MIN_FORWARD_NUMBER); // "나"가 우승자
-
-        assertThat(racingCarService.findWinners(cars)).containsAll(List.of("나"));
-    }
-
-    @DisplayName("우승자 여러명 테스트")
-    @Test
-    void findWinnersTest() {
-        Cars cars = new Cars(List.of("가", "나", "다"));
-
-        cars.getCars().get(0).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(0).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(1).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
-
-        assertThat(racingCarService.findWinners(cars)).containsAll(List.of("가", "다"));
-    }
-
-    @DisplayName("모두가 우승자 테스트")
-    @Test
-    void findWinnersAllTest() {
-        Cars cars = new Cars(List.of("가", "나", "다"));
-
-        cars.getCars().get(0).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(0).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(1).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(1).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
-        cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
-
-        assertThat(racingCarService.findWinners(cars)).containsAll(List.of("가", "나", "다"));
-    }
-
-
 }
